+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to the "jenkinsfile-support" extension will be documented in
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## [Unreleased]
-### Empty
-
+### Added
+- LSP - on hover action
+- LSP - auto completion
 ## [0.2.0] - 2018-10-18
 ### Added
 - multiline comments